jagua_rs::collision_detection::hazards

Enum BasicHazardEntity

pub enum BasicHazardEntity {
    PlacedItem {
        id: usize,
        dt: DTransformation,
        pk: PItemKey,
    },
    Exterior,
    Hole {
        idx: usize,
    },
    InferiorQualityZone {
        quality: usize,
        idx: usize,
    },
}

Entity inducing a Hazard in this crate’s layouts. All entities are uniquely identified.

Variants

PlacedItem

An item placed in the layout, defined by its id, applied transformation and key

Fields

id: usize

dt: DTransformation

pk: PItemKey

Exterior

Represents all regions outside the container

Hole

Represents a hole in the container.

Fields

idx: usize

InferiorQualityZone

Represents a zone in the container with a specific quality level that is inferior to the base quality.

Fields

quality: usize

idx: usize