jagua_rs::collision_detection::hazards

Module filter

Everything related to hazard filters

Structs

HazKeyFilter

Deems hazards with specific HazKey’s as irrelevant.

NoFilter

Deems no hazards as irrelevant.

Traits

HazardFilter

Trait for filters to ignore all Hazards induced by specific HazardEntitys. Enables collision queries to ignore specific hazards during the check.