jagua_rs::collision_detection::hazards::filter

Trait HazardFilter

pub trait HazardFilter {
    // Required method
    fn is_irrelevant(&self, haz_key: HazKey) -> bool;
}

Trait for filters to ignore all Hazards induced by specific HazardEntitys. Enables collision queries to ignore specific hazards during the check.

Required Methods

fn is_irrelevant(&self, haz_key: HazKey) -> bool

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".