Trait jagua_rs::collision_detection::quadtree::qt_traits::QTQueryable

pub trait QTQueryable: Shape + CollidesWith<Edge> + CollidesWith<AARectangle> { }

Common trait for all geometric primitives that can be directly queried in the quadtree for collisions with the edges of the registered hazards. These include: AARectangle, Edge and Circle.