jagua_rs::collision_detection::quadtree

Trait QTQueryable

pub trait QTQueryable: CollidesWith<Edge> + CollidesWith<Rect> { }

Common trait for all geometric primitives that can be directly queried in the quadtree for collisions with the edges of the registered hazards. These include: Rect, Edge and Circle.