jagua_rs

Module entities

Entities to model 2D Irregular Cutting and Packing Problems

Structs

Container

A container in which Item’s can be placed.

InferiorQualityZone

Represents a zone of inferior quality in the Container

Item

Item to be produced.

Layout

A Layout is a dynamic representation of items that have been placed in a container at specific positions. Items can be placed and removed. The container can be swapped. Snapshots can be taken and restored to. Each layout maintains a CDEngine, which can be used to check for collisions before placing items.

LayoutSnapshot

Immutable and compact representation of a Layout. Can be used to restore a Layout back to a previous state.

PItemKey

Unique key for each PlacedItem in a layout.

PlacedItem

Represents an Item that has been placed in a Layout

Constants

N_QUALITIES

Maximum number of qualities that can be used for quality zones in a container.

Traits

Instance

The (abstract) static representation of a problem instance. This trait defines shared functionality between any instance variant.