Enum jagua_rs::entities::problems::problem::Problem

pub enum Problem {
    BP(BPProblem),
    SP(SPProblem),
}

Enum which contains all the different problem types. A Problem represents a problem instance in a modifiable state. It can insert or remove items, create a snapshot from the current state called a Solution, and restore its state to a previous Solution.
Also enables the use of match statements on the Problem enum when variant-specific behavior is required, When a new variant is added, compile errors will be generated everywhere specific behaviour is required.

Variants

BP(BPProblem)

Bin Packing Problem

SP(SPProblem)

Strip Packing Problem