Trait jagua_rs::entities::problems::problem_generic::ProblemGeneric

pub trait ProblemGeneric: ProblemGenericPrivate {

    // Required methods
    fn place_item(&mut self, i_opt: &PlacingOption) -> LayoutIndex;
    fn remove_item(
        &mut self,
        layout_index: LayoutIndex,
        pi_uid: &PlacedItemUID,
        commit_instantly: bool
    );
    fn create_solution(&mut self, old_solution: &Option<Solution>) -> Solution;
    fn restore_to_solution(&mut self, solution: &Solution);
    fn layouts(&self) -> &[Layout];
    fn layouts_mut(&mut self) -> &mut [Layout];
    fn template_layouts(&self) -> &[Layout];
    fn missing_item_qtys(&self) -> &[isize];
    fn bin_qtys(&self) -> &[usize];
    fn instance(&self) -> &dyn InstanceGeneric;

    // Provided methods
    fn placed_item_qtys(&self) -> impl Iterator<Item = usize> { ... }
    fn usage(&mut self) -> fsize { ... }
    fn used_bin_cost(&self) -> u64 { ... }
    fn layout_indices(&self) -> impl Iterator<Item = LayoutIndex> { ... }
    fn template_layout_indices_with_stock(
        &self
    ) -> impl Iterator<Item = LayoutIndex> { ... }
    fn get_layout(&self, index: impl Borrow<LayoutIndex>) -> &Layout { ... }
    fn flush_changes(&mut self) { ... }

}

Trait for public shared functionality of all problem variants.

Required Methods

fn place_item(&mut self, i_opt: &PlacingOption) -> LayoutIndex

Places an item into the problem instance according to the given PlacingOption. Returns the index of the layout where the item was placed.

fn remove_item( &mut self, layout_index: LayoutIndex, pi_uid: &PlacedItemUID, commit_instantly: bool )

Removes an item with a specific PlacedItemUID from a specific Layout For more information about commit_instantly, see crate::collision_detection::cd_engine::CDEngine::deregister_hazard.

fn create_solution(&mut self, old_solution: &Option<Solution>) -> Solution

Saves the current state of the problem as a Solution.

fn restore_to_solution(&mut self, solution: &Solution)

Restores the state of the problem to a previous Solution.

fn layouts(&self) -> &[Layout]

fn layouts_mut(&mut self) -> &mut [Layout]

fn template_layouts(&self) -> &[Layout]

Template layouts are empty and immutable. For every unique bin in the problem instance, there is a template layout. When an item is placed in a template layout, it is cloned into a real layout.

fn missing_item_qtys(&self) -> &[isize]

The quantity of each item that is requested but currently missing in the problem instance, indexed by item id.

fn bin_qtys(&self) -> &[usize]

fn instance(&self) -> &dyn InstanceGeneric

Provided Methods

fn placed_item_qtys(&self) -> impl Iterator<Item = usize>

The quantity of each item that is currently placed in the problem instance, indexed by item id.

fn usage(&mut self) -> fsize

fn used_bin_cost(&self) -> u64

fn layout_indices(&self) -> impl Iterator<Item = LayoutIndex>

Returns the LayoutIndex of all layouts.

fn template_layout_indices_with_stock( &self ) -> impl Iterator<Item = LayoutIndex>

Returns the LayoutIndex of all template layouts that have remaining stock.

fn get_layout(&self, index: impl Borrow<LayoutIndex>) -> &Layout

fn flush_changes(&mut self)

Object Safety

This trait is not object safe.