jagua_rs::entities

Struct Container

pub struct Container {
    pub id: usize,
    pub outer_orig: Arc<OriginalShape>,
    pub outer_cd: Arc<SPolygon>,
    pub quality_zones: [Option<InferiorQualityZone>; 10],
    pub base_cde: Arc<CDEngine>,
}

A container in which Itemโs can be placed.

Fields
id: usize

Unique identifier of the container

outer_orig: Arc<OriginalShape>

Original contour of the container as defined in the input

outer_cd: Arc<SPolygon>

Contour of the container to be used for collision detection

quality_zones: [Option<InferiorQualityZone>; 10]

Zones of different qualities in the container, stored per quality.

base_cde: Arc<CDEngine>

The initial state of the CDEngine for this container. (equivalent to an empty layout using this container)

Implementations

impl Container

pub fn new( id: usize, original_outer: OriginalShape, quality_zones: Vec<InferiorQualityZone>, cde_config: CDEConfig, ) -> Result<Self>

pub fn area(&self) -> f32

The area of the contour of the container, excluding holes