jagua_rs::geometry

Module fail_fast

The fail-fast surrogate and all logic pertaining to its generation

Structs

SPSurrogate

Surrogate representation of a SPolygon - a ‘light-weight’ representation that is fully contained in the original SPolygon. Used for fail-fast collision detection.

SPSurrogateConfig

Configuration of the SPSurrogate generation

Functions

compute_pole

Computes the pole - the largest circle which is both inside of shape while being outside all other poles. Closely related to Pole of Inaccessibility (PoI), and inspired by Mapbox’s polylabel algorithm.

generate_piers

Generates a set of n piers - line segments fully contained within shape. This function generates them in such a way as to cover areas of the shape that are poorly represented by poles as well as possible.

generate_surrogate_poles

Generates a set of ‘poles’ for a shape according to specified coverage limits. See compute_pole for details on what a ‘pole’ is.