Function jagua_rs::geometry::fail_fast::poi::generate_next_pole

pub fn generate_next_pole(shape: &SimplePolygon, poles: &[Circle]) -> Circle

Generates the Pole of Inaccessibility (PoI). The PoI is the point in the interior of the shape that is farthest from the boundary. The interior is defined as the interior of the shape minus the interior of the poles.