jagua_rs::geometry

Module geo_traits

Set of traits representing various geometric properties & operations

Traits

AlmostCollidesWith

Trait for types that can detect ‘almost-collisions’ between self and other of type T.

CollidesWith

Trait for types that can detect collisions between self and other of type T.

DistanceTo

Trait for types that can compute the minimum distance between self and other of type T.

SeparationDistance

Trait for types that can compute the minimum distance to separate self from other of type T.

Transformable

Trait for types that can modify themselves by a Transformation.

TransformableFrom

Trait for types that can modify themselves to a reference object with a Transformation applied.