Trait jagua_rs::geometry::geo_traits::AlmostCollidesWith

pub trait AlmostCollidesWith<T> {
    // Required method
    fn almost_collides_with(&self, other: &T) -> bool;
}

Trait for types that can detect almost-collisions between itself and an object from type T. Useful in situations where fp arithmetic precision could be problematic. Should be implemented to lean towards false positives rather than false negatives.

Required Methods

fn almost_collides_with(&self, other: &T) -> bool