Trait jagua_rs::geometry::geo_traits::DistanceFrom

pub trait DistanceFrom<T> {
    // Required methods
    fn sq_distance(&self, other: &T) -> fsize;
    fn distance(&self, other: &T) -> fsize;
    fn distance_from_border(&self, other: &T) -> (GeoPosition, fsize);
    fn sq_distance_from_border(&self, other: &T) -> (GeoPosition, fsize);
}

Trait for geometric primitives that can calculate distances to other primitives.

Required Methods

fn sq_distance(&self, other: &T) -> fsize

Squared distance between the two primitives.

fn distance(&self, other: &T) -> fsize

Distance between the two primitives.

fn distance_from_border(&self, other: &T) -> (GeoPosition, fsize)

Distance from other to the boundary of self and whether other is in the interior or exterior of self.

fn sq_distance_from_border(&self, other: &T) -> (GeoPosition, fsize)

Squared distance from other to the boundary of self and whether other is in the interior or exterior of self.