Trait jagua_rs::geometry::geo_traits::TransformableFrom

pub trait TransformableFrom: Transformable {
    // Required method
    fn transform_from(
        &mut self,
        reference: &Self,
        t: &Transformation,
    ) -> &mut Self;
}

Trait for types that can be transformed based on a reference object with a Transformation applied. Used for repeated transformations on an identical reference shape without reallocating new memory each time.

Required Methods

fn transform_from(&mut self, reference: &Self, t: &Transformation) -> &mut Self

Object Safety

This trait is not object safe.