Struct jagua_rs::geometry::primitives::aa_rectangle::AARectangle

pub struct AARectangle {
    pub x_min: fsize,
    pub y_min: fsize,
    pub x_max: fsize,
    pub y_max: fsize,
}

Geometric primitive representing an axis-aligned rectangle

Fields
x_min: fsizey_min: fsizex_max: fsizey_max: fsize
Implementations

impl AARectangle

pub fn new(x_min: fsize, y_min: fsize, x_max: fsize, y_max: fsize) -> Self

pub fn top_edge(&self) -> Edge

pub fn bottom_edge(&self) -> Edge

pub fn left_edge(&self) -> Edge

pub fn right_edge(&self) -> Edge

pub fn edges(&self) -> [Edge; 4]

pub fn corners(&self) -> [Point; 4]

Returns the four corners in the following order: NW,NE,SW,SE

pub fn relation_to(&self, other: &AARectangle) -> GeoRelation

Returns the relation between self and another AARectangle

pub fn almost_relation_to(&self, other: &AARectangle) -> GeoRelation

Returns the relation between self and another AARectangle, with a tolerance for floating point precision. Leaning towards Surrounding and Enclosed instead of Intersecting in edge cases.

pub fn inflate_to_square(&self) -> AARectangle

Returns the rectangle that is the result of inflating the smallest dimension of the rectangle to match the largest dimension

pub fn scale(self, factor: fsize) -> Self

pub const QUADRANT_NEIGHBOR_LAYOUT: [[usize; 2]; 4] = _

For all quadrants, contains indices of the two neighbors of the quadrant at that index

pub fn quadrants(&self) -> [Self; 4]

Returns the 4 quadrants of the rectangle, in the order NW, NE, SW, SE

pub fn width(&self) -> fsize

pub fn height(&self) -> fsize

pub fn from_intersection( a: &AARectangle, b: &AARectangle, ) -> Option<AARectangle>

pub fn bounding_rectangle(a: &AARectangle, b: &AARectangle) -> AARectangle