jagua_rs::geometry::primitives

Struct Rect

pub struct Rect {
    pub x_min: f32,
    pub y_min: f32,
    pub x_max: f32,
    pub y_max: f32,
}

Axis-aligned rectangle

Fields
x_min: f32y_min: f32x_max: f32y_max: f32
Implementations

impl Rect

pub const QUADRANT_NEIGHBOR_LAYOUT: [[usize; 2]; 4]

For all quadrants, contains indices of the two neighbors of the quadrant at that index.

pub fn try_new(x_min: f32, y_min: f32, x_max: f32, y_max: f32) -> Result<Self>

pub fn from_diagonal_corners(c1: Point, c2: Point) -> Result<Self>

pub fn relation_to(&self, other: Rect) -> GeoRelation

Returns the geometric relation between self and another Rect. Optimized for GeoRelation::Disjoint

pub fn almost_relation_to(&self, other: Rect) -> GeoRelation

Returns the GeoRelation between self and another Rect, with a tolerance for floating point precision. In edge cases, this method will lean towards Surrounding and Enclosed instead of Intersecting.

pub fn inflate_to_square(&self) -> Rect

Returns a new rectangle with the same centroid but inflated to be the minimum square that contains self.

pub fn scale(self, factor: f32) -> Self

Returns a new rectangle with the same centroid but scaled by factor.

pub fn resize_by(self, dx: f32, dy: f32) -> Option<Self>

Returns a new rectangle with the same centroid as self but expanded by dx in both x-directions and by dy in both y-directions. If the new rectangle is invalid (x_min >= x_max or y_min >= y_max), returns None.

pub fn quadrants(&self) -> [Self; 4]

Returns the 4 quadrants of self. Ordered in the same way as quadrants in a cartesian plane: https://en.wikipedia.org/wiki/Quadrant_(plane_geometry)

pub fn corners(&self) -> [Point; 4]

Returns the four corners of self, in the same order as Rect::quadrants.

pub fn sides(&self) -> [Edge; 4]

Returns the four sides that make up self, in the same order as Rect::quadrants.

pub fn edges(&self) -> [Edge; 4]

Returns the four edges that make up self, in the same order as Rect::quadrants.

pub fn width(&self) -> f32

pub fn height(&self) -> f32

pub fn intersection(a: Rect, b: Rect) -> Option<Rect>

Returns the largest rectangle that is contained in both a and b.

pub fn bounding_rect(a: Rect, b: Rect) -> Rect

Returns the smallest rectangle that contains both a and b.

pub fn centroid(&self) -> Point

pub fn area(&self) -> f32

pub fn diameter(&self) -> f32