jagua_rs::geometry::primitives

Struct SPolygon

pub struct SPolygon {
    pub vertices: Vec<Point>,
    pub bbox: Rect,
    pub area: f32,
    pub diameter: f32,
    pub poi: Circle,
    pub surrogate: Option<SPSurrogate>,
}

A Simple Polygon is a polygon that does not intersect itself and contains no holes. It is a closed shape with a finite number of vertices and edges. read more

Fields
vertices: Vec<Point>

Set of points that form the polygon

bbox: Rect

Bounding box

area: f32

Area of its interior

diameter: f32

Maximum distance between any two points in the polygon

poi: Circle

Pole of inaccessibility represented as a circle

surrogate: Option<SPSurrogate>

Optional surrogate representation of the polygon (subset of the original)

Implementations

impl SPolygon

pub fn new(points: Vec<Point>) -> Result<Self>

Create a new simple polygon from a set of points, expensive operations are performed here! Use Self::clone() or Self::transform() to avoid recomputation.

pub fn generate_surrogate(&mut self, config: SPSurrogateConfig) -> Result<()>

pub fn vertex(&self, i: usize) -> Point

pub fn edge(&self, i: usize) -> Edge

pub fn edge_iter(&self) -> impl Iterator<Item = Edge> + '_

pub fn n_vertices(&self) -> usize

pub fn surrogate(&self) -> &SPSurrogate

pub fn calculate_diameter(points: Vec<Point>) -> f32

pub fn generate_bounding_box(points: &[Point]) -> Rect

pub fn calculate_area(points: &[Point]) -> f32

pub fn calculate_poi(points: &[Point], diameter: f32) -> Result<Circle>

pub fn centroid(&self) -> Point