jagua_rs::geometry

Module shape_modification

Set of function to modify geometric shapes

Structs

ShapeModifyConfig

Enums

ShapeModifyMode

Whether to strictly inflate or deflate when making any modifications to shape. Depends on the position of the HazardEntity that the shape represents.

Functions

close_narrow_concavities

Closes narrow concavities in a SPolygon by replacing them with a straight edge, eliminating the vertices in between.

offset_shape

Offsets a SPolygon by a certain distance either inwards or outwards depending on the ShapeModifyMode. Relies on the geo_offset crate.

shape_modification_valid

simplify_shape

Simplifies a SPolygon by reducing the number of edges.