jagua_rs::geometry::shape_modification

Function simplify_shape

pub fn simplify_shape(
    shape: &SPolygon,
    mode: ShapeModifyMode,
    max_area_change_ratio: f32,
) -> SPolygon

Simplifies a SPolygon by reducing the number of edges.

The simplified shape will either be a subset or a superset of the original shape, depending on the ShapeModifyMode. The procedure sequentially eliminates edges until either the change in area (ratio) exceeds max_area_delta or the number of edges < 4.