jagua_rs::io

Module ext_repr

External (serializable) representations of the entities within the library.

Structs

ExtContainer

External representation of a Container. Items can be placed inside containers.

ExtItem

External representation of an Item.

ExtLayout

External representation of a Layout. A layout consists of a container with items placed in a specific configuration.

ExtPlacedItem

External representation of a PlacedItem. An item placed in a container with a specific transformation.

ExtPolygon

A polygon represented as an outer boundary and a list of holes

ExtQualityZone

A zone with a specific quality level

ExtSPolygon

External representation of a SPolygon. A polygon with no holes and no self-intersections.

ExtTransformation

Represents a proper rigid transformation defined as a rotation followed by translation

Enums

ExtShape

Various ways to represent a shape