jagua_rs::probs::bpp

Module entities

Entities to model the Bin Packing Problem (BPP)

Structs

BPInstance

Instance of the Bin Packing Problem.

BPPlacement

Encapsulates all required information to place an Item in a BPProblem.

BPProblem

Dynamic counterpart of BPInstance.

BPSolution

Snapshot of BPProblem at a specific moment. Can be used to restore to a previous state.

Bin

A container in the Bin Packing Problem (BPP) with an associated cost and stock.

LayKey

Unique key for each Layout in a BPProblem and BPSolution

Enums

BPLayoutType

Enum to distinguish between already existing Layouts and new ones.