Crate jagua_rs

A fast and fearless Collision Detection Engine for 2D irregular cutting and packing problems.

This library is designed to be used as a backend by optimization algorithms.

spp — Enables support for the Strip Packing Problem

bpp — Enables support for the Bin Packing Problem

mspp — Enables support for the Multi Strip Packing Problem

Modules

collision_detection

Everything related to the Collision Detection Engine

entities

Entities to model 2D Irregular Cutting and Packing Problems

geometry

Geometric primitives and base algorithms

io

Importing problem instances into and exporting solutions out of this library

probs

Enabled variants of the 2D irregular Cutting and Packing Problem.

util

Helper functions which do not belong to any specific module

Type Aliases

Instant

Export the web_time crate’s Instant type for compatibility with web environments.