jagua_rs::collision_detection::hazards::filter

Trait HazardFilter

pub trait HazardFilter {
    // Required method
    fn is_irrelevant(&self, entity: &HazardEntity) -> bool;
}

Trait for filters to ignore all Hazards induced by specific HazardEntitys. Enables collision queries to ignore specific hazards during the check.

Required Methods

fn is_irrelevant(&self, entity: &HazardEntity) -> bool