jagua_rs::collision_detection::quadtree

Trait QTQueryable

pub trait QTQueryable: CollidesWith<Edge> + CollidesWith<Rect> {
    // Provided method
    fn collides_with_quadrants(&self, _r: &Rect, qs: [&Rect; 4]) -> [bool; 4] { ... }
}

Common trait for all geometric primitives that can be directly queried in the quadtree for collisions with the edges of the registered hazards. These include: Rect, Edge and Circle.

Provided Methods

fn collides_with_quadrants(&self, _r: &Rect, qs: [&Rect; 4]) -> [bool; 4]

Checks