jagua_rs::util

Module assertions

Set of functions used throughout assure the correctness of the library.

Functions

collision_hazards_sorted_correctly

layout_qt_matches_fresh_qt

print_layout

Prints code to rebuild a layout. Intended for debugging purposes.

qt_contains_no_dangling_hazards

quadrants_have_valid_layout

Checks if the quadrants follow the layout set in Rect::QUADRANT_NEIGHBOR_LAYOUT

snapshot_matches_layout